// src/entity/venda/venda.enums.ts
export enum CategoriaVenda {
  CONSULTORIA = 'Serviços Consultoria',
  LICENCAS = 'Licenças Software',
  SUPORTE = 'Suporte Técnico',
  CUSTOMIZADO = 'Desenvolvimento Customizado',
  IMPLANTACAO = 'Implantação Sistema',
  TREINAMENTO = 'Treinamento',
}

export enum StatusVenda {
  CONCLUIDA = 'Concluida',
  PENDENTE = 'Pendente',
  CANCELADA = 'Cancelada',
  PROCESSANDO = 'Processando',
}