// src/entity/reembolso/reembolso.enums.ts
export enum CategoriaReembolso {
  COMBUSTIVEL = 'Combustível',
  ALIMENTACAO = 'Alimentação',
  TRANSPORTE = 'Transporte',
  HOSPEDAGEM = 'Hospedagem',
  MATERIAL_ESCRITORIO = 'Material de Escritório',
  OUTROS = 'Outros',
}

export enum StatusReembolso {
  PENDENTE = 'Pendente',
  APROVADO = 'Aprovado',
  REJEITADO = 'Rejeitado',
  RASCUNHO = 'Rascunho',
}